"Works", in that it sends to both the right and left clients, but the game is not updating.
If you manage to fix that, I think we will
be basically done.

Plus, I don't know how, but it was able
to survive an entire game
While I was moving around the
paddle, so I hope that's a good sign
of things to come.

The changes I made were implemented primarily in pS1,
where I removed the multithreading I had earlier and just let the server
handle the traffic as needed.

The prints should be very helpful in debugging, they
definitely were that way for me.